refactor: extract LogicRulesEditor from ChecklistBuilderModal
Moves the ~280-line logic rules IIFE into a standalone LogicRulesEditor component,
reducing ChecklistBuilderModal by ~270 lines and making each file easier to load per session.
